Fix compiler warning
Signed unsigned mismatch.

+++ src/win32/cd_win.c
@@ -382,7 +382,7 @@ static void CD_f (void)
 
 LONG CDAudio_MessageHandler(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
 {
-	if (lParam != wDeviceID)
+	if ((UINT)lParam != wDeviceID)
 		return 1;
 
 	switch (wParam)
